+++ cli/05_train_pytorch_models/greedy/generate.py
@@ -20,7 +20,7 @@ def greedy_generate(
     generated = torch.full(
         (batch_size, 1),
         fill_value=bos_id,
-        dtype=torch.long,
+        dtype=input_ids.dtype,
         device=input_ids.device,
     )
     finished = torch.zeros(batch_size, dtype=torch.bool, device=input_ids.device)
